Update cachematrix.R
This is the answer for assignment 2 (week3) of the R programming course. Listed are the two functions: makeCacheMatrix, cacheSolve.
The first function creates a vector which is a list of functions to set the inverse of a special matrix object, while the second function is a function that calculates the inverse of a matrix set by the makeCacheMatrix function above.